Refactor profiler registration
