+++ src/fairseq2/assets/setup.py
@@ -17,7 +17,7 @@ from fairseq2.assets import (
     WheelPackageFileLister,
 )
 from fairseq2.context import RuntimeContext
-from fairseq2.setup._error import SetupError
+from fairseq2.error import SetupError
 from fairseq2.utils.yaml import StandardYamlLoader
