+++ src/fairseq2/cli/_main.py
@@ -14,7 +14,7 @@ from torch.cuda import OutOfMemoryError
 
 from fairseq2.cli._logging import setup_logging
 from fairseq2.context import get_runtime_context
-from fairseq2.error import ContractError, InternalError
+from fairseq2.error import ContractError, InternalError, SetupError
 from fairseq2.extensions import ExtensionError
 from fairseq2.logging import LoggingSetupError, log
 
@@ -46,7 +46,7 @@ def main() -> None:
 
 
 def _run() -> int:
-    from fairseq2.setup import SetupError, setup_cli, setup_fairseq2
+    from fairseq2.setup import setup_cli, setup_fairseq2
 
     try:
         setup_logging()
